create tests for chat API

+++ internal/api/chat/delete_chat.go
@@ -2,6 +2,7 @@ package chat
 
 import (
 	"context"
+	"fmt"
 
 	desc "github.com/merynayr/chat-server/pkg/chat_v1"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -9,6 +10,9 @@ import (
 
 // DeleteChat - отправляет запрос в сервисный слой на создание пользователя
 func (a *API) DeleteChat(ctx context.Context, req *desc.DeleteChatRequest) (*emptypb.Empty, error) {
+	if req == nil {
+		return nil, fmt.Errorf("Request is nil")
+	}
 	err := a.chatService.DeleteChat(ctx, req.GetId())
 	if err != nil {
 		return nil, err
